Rename: 메인페이지 feature컴포넌트 폴더 생성 및 이동
