+++ src/components/feature/main/PostList.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { useQuery } from 'react-query';
-import { PostCardData } from '../../types/commonTypes';
+import { PostCardData } from '../../../types/commonTypes';
 import PostItem from './PostItem';
 import styled from 'styled-components';
 import SkeletonPost from './SkeletonPost';
-import api from '../../services/api';
+import api from '../../../services/api';
 
 const PostList = () => {
   const { isLoading, data } = useQuery<{ posts: PostCardData[]; isLastPage: boolean }>('posts', api.getPost);
